refactor(113181): cria pasta para modal
